Fix crop keyframe validation error by using float precision
Changes:
- Update CropKeyframe model to accept floats instead of integers
- Add round3() function to round values to 3 decimal places
- Apply rounding in CropOverlay: constrainCrop, applyAspectRatio, handleMouseUp
- Apply rounding in useCrop: interpolateCrop
- Maintain precision during transit while preventing validation errors

This ensures crop values maintain 3 decimal point precision throughout
the system, preventing Pydantic validation errors while keeping values
in sync between frontend and backend.

Fixes #export-validation-error

+++ src/backend/app/main.py
@@ -121,10 +121,10 @@ async def get_status():
 # Crop Export Models
 class CropKeyframe(BaseModel):
     time: float
-    x: int
-    y: int
-    width: int
-    height: int
+    x: float
+    y: float
+    width: float
+    height: float
 
 
 class CropExportRequest(BaseModel):
